Error Pictures
I couldn't get the app to deploy on my local server. I set up a firewall for port 8501 in GCP and tried all practices to get it on my local server. I was able to execute all python code though and I was able to get "streamlit" to work through GCP